% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingvalues.R
\name{count_all_missing_by_group}
\alias{count_all_missing_by_group}
\title{To count the missing values for all columns by group}
\usage{
count_all_missing_by_group(data, group_col)
}
\arguments{
\item{data}{Data in data.frame or tibble format}

\item{group_col}{Column in `data` to group by (as a string)}
}
\value{
A data frame with the levels of `group`, and the number of NAs
  within each level of `group` for all columns in `data` except `group`.
}
\description{
Given a data frame `data` and a column `group`, this function creates
a new data frame with one row per level of `group`. The first column
contains the levels of `group`, and the rest of the columns contain
the number of missing values for all columns in `data` except `group`.
}
\examples{
# Example 1: Count missing values by Month
count_all_missing_by_group(airquality, "Month")

# Example 2: Count missing values by Wind (grouped by wind speeds)
# For this example, we first create a new column to categorize `Wind` speeds
airquality$Wind_category <- cut(airquality$Wind, breaks=c(0, 5, 10, 15, 20),
                                labels=c("Low", "Medium", "High", "Very High"))
count_all_missing_by_group(airquality, "Wind_category")

# Example 3: Count missing values by Temp (grouped by temperature ranges)
# For this example, we categorize `Temp` values
airquality$Temp_category <- cut(airquality$Temp, breaks=c(50, 70, 80, 90, 100),
                                labels=c("Cool", "Moderate", "Warm", "Hot"))
count_all_missing_by_group(airquality, "Temp_category")

}
